Complete first pass on testing system.  All tests pass on my box.  Whew.
We now have a first cut of a functioning testing system.  Many
examples had to be disabled (with a proper decorator), but now we can
begin building a robust test suite for all of IPython.

The big remaining todo is to fix the handling of the ipython/test
namespace synchronization, which is still not 100% correct.  Once that
is done, some of the currently skipped tests will be reactivated.